pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mapkureader"
version = "0.1.0"
description = "Machines Reading Maps — text extraction, visual classification, and georeferencing for historical map images"
readme = "README.md"
license = "Unlicense"
requires-python = ">=3.10"
authors = [
    {name = "Jon Jablonski", email = "jonjab@ucsb.edu"},
]

dependencies = [
    "numpy>=1.24",
    "Pillow>=10.0",
    "rasterio>=1.3",
    "shapely>=2.0",
    "pyproj>=3.6",
    "geopandas>=0.14",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
text = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.30",
]
classify = [
    "torch>=2.0",
    "torchvision>=0.15",
    "scikit-learn>=1.3",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
